% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_day_indices.R
\name{define_day_indices}
\alias{define_day_indices}
\title{Define day indices from timestamp}
\usage{
define_day_indices(ts)
}
\arguments{
\item{ts}{Character (no default) with timestamps from data.}
}
\value{
Numeric vector with day indices
}
\description{
Define day indices from timestamp
}
\examples{
\dontrun{
data = readFile("mydata/s01_fitbit.csv")
define_day_indices(data$timestamp)
}
}
