% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cadence_bands.R
\name{get_cadence_bands}
\alias{get_cadence_bands}
\title{Function to calculate cadence bands}
\usage{
get_cadence_bands(x, bands = c(0, 1, 20, 40, 60, 80, 100, 120, Inf))
}
\arguments{
\item{x}{Numeric vector (no default) with cadence data for the day of interest.}

\item{bands}{Numeric vector (default = c(0, 1, 20, 40, 60, 80, 100, 120, Inf)) with the bands of interest to calculate.}
}
\value{
Cadence bands and names of variables.
}
\description{
Function to calculate cadence bands
}
\examples{
\dontrun{
get_cadence_bands(x, peaks = c(1, 30, 60))
}
}
