% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cadence_peaks.R
\name{get_cadence_peaks}
\alias{get_cadence_peaks}
\title{Function to calculate cadence peaks}
\usage{
get_cadence_peaks(x, peaks = c(1, 30, 60))
}
\arguments{
\item{x}{Numeric vector (no default) with cadence data for the day of interest.}

\item{peaks}{Numeric vector (default = c(1, 30, 60)) with the peaks of interest to calculate.}
}
\value{
Cadence peaks, number of zeroes in each peak, names of variables.
}
\description{
Function to calculate cadence peaks
}
\examples{
\dontrun{
get_cadence_peaks(x, peaks = c(1, 30, 60))
}
}
