% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isGGIRoutput.R
\name{isGGIRoutput}
\alias{isGGIRoutput}
\title{Checks whether a directory is a GGIR output directory}
\usage{
isGGIRoutput(path)
}
\arguments{
\item{path}{Character (no default). Directory to test if it is a GGIR output directory}
}
\value{
Boolean: TRUE means it is GGIR output.
}
\description{
Checks whether a directory is a GGIR output directory
}
\examples{
\dontrun{
isGGIRoutput("C:/mystudy/output_GGIR/")
}
}
