% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFile.R
\name{readFile}
\alias{readFile}
\title{Reads and formats the timestamp and step data for one participant.}
\usage{
readFile(path, time_format = c())
}
\arguments{
\item{path}{Character (no default). Path to file or files with the timestamp and step data for one participant.}

\item{time_format}{Character (no default, optional). String indicating the format of the timestamp in data (as in strptime)}
}
\value{
Data frame containing timestamp and steps in a clean format for one participant.
}
\description{
Reads and formats the timestamp and step data for one participant.
}
\examples{
\dontrun{
readFile(path)
}
}
