% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.metrics.R
\name{step.metrics}
\alias{step.metrics}
\title{Function to calculate and export step metrics.}
\usage{
step.metrics(
  datadir,
  outputdir = "./",
  id_sep = "_",
  threshold.MOD = 100,
  threshold.VIG = 130,
  includedaycrit = 10,
  exclude_pk30_0 = TRUE,
  exclude_pk60_0 = TRUE,
  time_format = NULL
)
}
\arguments{
\item{datadir}{Character (no default). Directory where the files containing the steps data are stored, e.g., "C:/mydata/".}

\item{outputdir}{Character (no default). Directory where the output needs to be stored. Note that this function will attempt to create folders in this directory and uses those folder to keep output.}

\item{threshold.MOD}{Numeric (default = 100)}

\item{threshold.VIG}{Numeric (default = 130)}

\item{includedaycrit}{Numeric (default = 10)}

\item{exclude_pk30_0}{Logical (default = TRUE)}

\item{exclude_pk60_0}{Logical (default = TRUE)}

\item{time_format}{Character (default = NULL)}
}
\value{
This function does not return any object. It stores csv files with the
         day-level and person-level data in the output directory.
}
\description{
This function loads the minute-based files containing the step counts,
               calculate the step and cadence metrics and store them in day-level and person-level csv files.
}
\examples{
\dontrun{
step.metrics(datadir = "C:/mydata/", outputdir = "D:/myoutput/")
}
}
